Add daily newsletter, change instant notifications
A daily newsletter about content changes and new comments is sent.
This is done by Flags, Views Bulk Operations and Rules. The news-
letter is scheduled to run daily at 14:00 (2:00 pm). Members can
opt-out on their account page. The updates are sent only about the
content that the member has subscribed to using the Follow links.
There is a overview page of all subscriptions at
user/%/cmtls/subscriptions.

The instant notifications can now also be opted-out on the account
page.